Towards WorldRenderContext
