+++ Game/RenderContext.h
@@ -16,6 +16,7 @@
 #include "TextureTypes.h"
 #include "UploadedTextureManager.h"
 #include "ViewModel.h"
+#include "WorldRenderContext.h"
 
 #include <GameOpenGL/GameOpenGL.h>
 #include <GameOpenGL/GameOpenGLMappedBuffer.h>
@@ -176,9 +177,14 @@ public:
 
     void SetAmbientLightIntensity(float intensity)
     {
+        // Assume calls are already damped
+
         mAmbientLightIntensity = intensity;
 
-        mRenderParameters.EffectiveAmbientLightIntensity = CalculateEffectiveAmbientLightIntensity();
+        // Re-calculate effective ambient light intensity
+        mRenderParameters.EffectiveAmbientLightIntensity = CalculateEffectiveAmbientLightIntensity(
+            mAmbientLightIntensity,
+            mWorldRenderContext->GetStormAmbientDarkening());
         mRenderParameters.IsEffectiveAmbientLightIntensityDirty = true;
     }
 
@@ -536,43 +542,48 @@ public:
 
     void UploadStart();
 
-    void UploadStarsStart(size_t starCount);
+    inline void UploadStarsStart(size_t starCount)
+    {
+        mWorldRenderContext->UploadStarsStart(starCount);
+    }
 
     inline void UploadStar(
         float ndcX,
         float ndcY,
         float brightness)
     {
-        //
-        // Populate vertex in buffer
-        //
-
-        mStarVertexBuffer.emplace_back(ndcX, ndcY, brightness);
+        mWorldRenderContext->UploadStar(
+            ndcX,
+            ndcY,
+            brightness);
     }
 
-    void UploadStarsEnd();
+    inline void UploadStarsEnd()
+    {
+        mWorldRenderContext->UploadStarsEnd();
+    }
 
     inline void UploadStormAmbientDarkening(float darkening)
-    {
-        if (darkening != mStormAmbientDarkening)
+    {        
+        if (mWorldRenderContext->UploadStormAmbientDarkening(darkening))
         {
-            mStormAmbientDarkening = darkening;
-
-            mRenderParameters.EffectiveAmbientLightIntensity = CalculateEffectiveAmbientLightIntensity();
+            mRenderParameters.EffectiveAmbientLightIntensity = CalculateEffectiveAmbientLightIntensity(
+                mAmbientLightIntensity,
+                mWorldRenderContext->GetStormAmbientDarkening());
+            
             mRenderParameters.IsEffectiveAmbientLightIntensityDirty = true;
         }
     }
 
     inline void UploadRain(float density)
     {
-        if (density != mRainDensity)
-        {
-            mRainDensity = density;
-            mIsRainDensityDirty = true;
-        }
+        mWorldRenderContext->UploadRain(density);
     }
 
-    void UploadLightningsStart(size_t lightningCount);
+    inline void UploadLightningsStart(size_t lightningCount)
+    {
+        mWorldRenderContext->UploadLightningsStart(lightningCount);
+    }
 
     inline void UploadBackgroundLightning(
         float ndcX,
@@ -580,19 +591,12 @@ public:
         float renderProgress,
         float personalitySeed)
     {
-        // Get NDC coordinates of world y=0 (i.e. sea level)
-        float const ndcSeaLevel = mRenderParameters.View.WorldToNdc(vec2f::zero()).y;
-
-        // Store vertices
-        StoreLightningVertices(
+        mWorldRenderContext->UploadBackgroundLightning(
             ndcX,
-            ndcSeaLevel,
             progress,
             renderProgress,
             personalitySeed,
-            mBackgroundLightningVertexCount);
-
-        mBackgroundLightningVertexCount += 6;
+            mRenderParameters);
     }
 
     inline void UploadForegroundLightning(
@@ -601,27 +605,23 @@ public:
         float renderProgress,
         float personalitySeed)
     {
-        // Get NDC coordinates of tip point, a few metres down,
-        // to make sure tip touches visually the point
-        vec2f const ndcTip = mRenderParameters.View.WorldToNdc(
-            tipWorldCoordinates
-            + vec2f(0.0f, -3.0f));
-
-        // Store vertices
-        StoreLightningVertices(
-            ndcTip.x,
-            ndcTip.y,
+        mWorldRenderContext->UploadForegroundLightning(
+            tipWorldCoordinates,
             progress,
             renderProgress,
             personalitySeed,
-            mLightningVertexBuffer.max_size() - (mForegroundLightningVertexCount + 6));
-
-        mForegroundLightningVertexCount += 6;
+            mRenderParameters);
     }
 
-    void UploadLightningsEnd();
+    inline void UploadLightningsEnd()
+    {
+        mWorldRenderContext->UploadLightningsEnd();
+    }
 
-    void UploadCloudsStart(size_t cloudCount);
+    inline void UploadCloudsStart(size_t cloudCount)
+    {
+        mWorldRenderContext->UploadCloudsStart(cloudCount);
+    }
 
     inline void UploadCloud(
         uint32_t cloudId,
@@ -630,276 +630,85 @@ public:
         float scale,
         float darkening)    // 0.0:dark, 1.0:light
     {
+        mWorldRenderContext->UploadCloud(
+            cloudId,
+            virtualX,
+            virtualY,
+            scale,
+            darkening,
+            mRenderParameters);
         //
-        // We use Normalized Device Coordinates here
-        //
-
-        //
-        // Map input slice [-0.5, +0.5], [-0.5, +0.5] into NDC [-1.0, +1.0], [-1.0, +1.0]
-        //
-
-        float const mappedX = virtualX * 2.0f;
-        float const mappedY = virtualY * 2.0f;
-
-        // TEST CODE: this code fits everything in the visible window
-        ////float const mappedX = virtualX / 1.5f;
-        ////float const mappedY = virtualY / 1.5f;
-        ////scale /= 1.5f;
-
-
-        //
-        // Populate quad in buffer
-        //
-
-        float const aspectRatio = mRenderParameters.View.GetAspectRatio();
-
-        size_t const cloudTextureIndex = static_cast<size_t>(cloudId) % mCloudTextureAtlasMetadata->GetFrameMetadata().size();
-
-        auto cloudAtlasFrameMetadata = mCloudTextureAtlasMetadata->GetFrameMetadata(
-            CloudTextureGroups::Cloud,
-            static_cast<TextureFrameIndex>(cloudTextureIndex));
-
-        float leftX = mappedX - scale * cloudAtlasFrameMetadata.FrameMetadata.AnchorWorldX;
-        float rightX = mappedX + scale * (cloudAtlasFrameMetadata.FrameMetadata.WorldWidth - cloudAtlasFrameMetadata.FrameMetadata.AnchorWorldX);
-        float topY = mappedY + scale * (cloudAtlasFrameMetadata.FrameMetadata.WorldHeight - cloudAtlasFrameMetadata.FrameMetadata.AnchorWorldY) * aspectRatio;
-        float bottomY = mappedY - scale * cloudAtlasFrameMetadata.FrameMetadata.AnchorWorldY * aspectRatio;
-
-        // top-left
-        mCloudVertexBuffer.emplace_back(
-            leftX,
-            topY,
-            cloudAtlasFrameMetadata.TextureCoordinatesBottomLeft.x,
-            cloudAtlasFrameMetadata.TextureCoordinatesTopRight.y,
-            darkening);
-
-        // bottom-left
-        mCloudVertexBuffer.emplace_back(
-            leftX,
-            bottomY,
-            cloudAtlasFrameMetadata.TextureCoordinatesBottomLeft.x,
-            cloudAtlasFrameMetadata.TextureCoordinatesBottomLeft.y,
-            darkening);
-
-        // top-right
-        mCloudVertexBuffer.emplace_back(
-            rightX,
-            topY,
-            cloudAtlasFrameMetadata.TextureCoordinatesTopRight.x,
-            cloudAtlasFrameMetadata.TextureCoordinatesTopRight.y,
-            darkening);
-
-        // bottom-left
-        mCloudVertexBuffer.emplace_back(
-            leftX,
-            bottomY,
-            cloudAtlasFrameMetadata.TextureCoordinatesBottomLeft.x,
-            cloudAtlasFrameMetadata.TextureCoordinatesBottomLeft.y,
-            darkening);
-
-        // top-right
-        mCloudVertexBuffer.emplace_back(
-            rightX,
-            topY,
-            cloudAtlasFrameMetadata.TextureCoordinatesTopRight.x,
-            cloudAtlasFrameMetadata.TextureCoordinatesTopRight.y,
-            darkening);
-
-        // bottom-right
-        mCloudVertexBuffer.emplace_back(
-            rightX,
-            bottomY,
-            cloudAtlasFrameMetadata.TextureCoordinatesTopRight.x,
-            cloudAtlasFrameMetadata.TextureCoordinatesBottomLeft.y,
-            darkening);
     }
 
-    void UploadCloudsEnd();
+    inline void UploadCloudsEnd()
+    {
+        mWorldRenderContext->UploadCloudsEnd();
+    }
 
-    void UploadLandStart(size_t slices);
+    inline void UploadLandStart(size_t slices)
+    {
+        mWorldRenderContext->UploadLandStart(slices);
+    }
 
     inline void UploadLand(
         float x,
         float yLand)
     {
-        float const yVisibleWorldBottom = mRenderParameters.View.GetVisibleWorldBottomRight().y;
-
-        //
-        // Store Land element
-        //
-
-        LandSegment & landSegment = mLandSegmentBuffer.emplace_back();
-
-        landSegment.x1 = x;
-        landSegment.y1 = yLand;
-        landSegment.depth1 = 0.0f;
-        landSegment.x2 = x;
-        // If land is invisible (below), then keep both points at same height, or else interpolated lines
-        // will have a slope varying with the y of the visible world bottom
-        float yBottom = yLand >= yVisibleWorldBottom ? yVisibleWorldBottom : yLand;
-        landSegment.y2 = yBottom;
-        landSegment.depth2 = -(yBottom - yLand); // Height of land
+        mWorldRenderContext->UploadLand(
+            x,
+            yLand,
+            mRenderParameters);
     }
 
-    void UploadLandEnd();
+    inline void UploadLandEnd()
+    {
+        mWorldRenderContext->UploadLandEnd();
+    }
 
-    void UploadOceanStart(size_t slices);
+    inline void UploadOceanStart(size_t slices)
+    {
+        mWorldRenderContext->UploadOceanStart(slices);
+    }
 
     inline void UploadOcean(
         float x,
         float yOcean,
         float oceanDepth)
     {
-        float const yVisibleWorldBottom = mRenderParameters.View.GetVisibleWorldBottomRight().y;
-
-        //
-        // Store ocean element
-        //
-
-        OceanSegment & oceanSegment = mOceanSegmentBuffer.emplace_back();
-
-        oceanSegment.x1 = x;
-        float const oceanSegmentY1 = yOcean;
-        oceanSegment.y1 = oceanSegmentY1;
-
-        oceanSegment.x2 = x;
-        float const oceanSegmentY2 = yVisibleWorldBottom;
-        oceanSegment.y2 = oceanSegmentY2;
-
-        switch (mRenderParameters.OceanRenderMode)
-        {
-            case OceanRenderModeType::Texture:
-            {
-                // Texture sample Y levels: anchor texture at top of wave,
-                // and set bottom at total visible height (after all, ocean texture repeats)
-                oceanSegment.value1 = 0.0f; // This is at yOcean
-                oceanSegment.value2 = yOcean - yVisibleWorldBottom; // Negative if yOcean invisible, but then who cares
-
-                break;
-            }
-
-            case OceanRenderModeType::Depth:
-            {
-                // Depth: top=0.0, bottom=height as fraction of ocean depth
-                oceanSegment.value1 = 0.0f;
-                oceanSegment.value2 = oceanDepth != 0.0f
-                    ? abs(oceanSegmentY2 - oceanSegmentY1) / oceanDepth
-                    : 0.0f;
-
-                break;
-            }
-
-            case OceanRenderModeType::Flat:
-            {
-                // Nop, but be nice
-                oceanSegment.value1 = 0.0f;
-                oceanSegment.value2 = 0.0f;
-
-                break;
-            }
-        }
+        mWorldRenderContext->UploadOcean(
+            x,
+            yOcean,
+            oceanDepth,
+            mRenderParameters);
     }
 
-    void UploadOceanEnd();
+    inline void UploadOceanEnd()
+    {
+        mWorldRenderContext->UploadOceanEnd();
+    }
 
-    void UploadAMBombPreImplosion(
+    inline void UploadAMBombPreImplosion(
         vec2f const & centerPosition,
         float progress,
         float radius)
     {
-        float const leftX = centerPosition.x - radius;
-        float const rightX = centerPosition.x + radius;
-        float const topY = centerPosition.y + radius;
-        float const bottomY = centerPosition.y - radius;
-
-        // Triangle 1
-
-        // Left, bottom
-        mAMBombPreImplosionVertexBuffer.emplace_back(
-            vec2f(leftX, bottomY),
-            centerPosition,
-            progress,
-            radius);
-
-        // Left top
-        mAMBombPreImplosionVertexBuffer.emplace_back(
-            vec2f(leftX, topY),
-            centerPosition,
-            progress,
-            radius);
-
-        // Right bottom
-        mAMBombPreImplosionVertexBuffer.emplace_back(
-            vec2f(rightX, bottomY),
-            centerPosition,
-            progress,
-            radius);
-
-        // Triangle 2
-
-        // Left top
-        mAMBombPreImplosionVertexBuffer.emplace_back(
-            vec2f(leftX, topY),
-            centerPosition,
-            progress,
-            radius);
-
-        // Right bottom
-        mAMBombPreImplosionVertexBuffer.emplace_back(
-            vec2f(rightX, bottomY),
-            centerPosition,
-            progress,
-            radius);
-
-        // Right, top
-        mAMBombPreImplosionVertexBuffer.emplace_back(
-            vec2f(rightX, topY),
+        mWorldRenderContext->UploadAMBombPreImplosion(
             centerPosition,
             progress,
             radius);
     }
 
-    void UploadCrossOfLight(
+    inline void UploadCrossOfLight(
         vec2f const & centerPosition,
         float progress)
     {
-        auto const & viewModel = mRenderParameters.View;
-
-        // Triangle 1
-
-        mCrossOfLightVertexBuffer.emplace_back(
-            vec2f(viewModel.GetVisibleWorldTopLeft().x, viewModel.GetVisibleWorldBottomRight().y), // left, bottom
+        mWorldRenderContext->UploadCrossOfLight(
             centerPosition,
-            progress);
-
-        mCrossOfLightVertexBuffer.emplace_back(
-            viewModel.GetVisibleWorldTopLeft(), // left, top
-            centerPosition,
-            progress);
-
-        mCrossOfLightVertexBuffer.emplace_back(
-            viewModel.GetVisibleWorldBottomRight(), // right, bottom
-            centerPosition,
-            progress);
-
-        // Triangle 2
-
-        mCrossOfLightVertexBuffer.emplace_back(
-            viewModel.GetVisibleWorldTopLeft(), // left, top
-            centerPosition,
-            progress);
-
-        mCrossOfLightVertexBuffer.emplace_back(
-            viewModel.GetVisibleWorldBottomRight(), // right, bottom
-            centerPosition,
-            progress);
-
-        mCrossOfLightVertexBuffer.emplace_back(
-            vec2f(viewModel.GetVisibleWorldBottomRight().x, viewModel.GetVisibleWorldTopLeft().y),  // right, top
-            centerPosition,
-            progress);
+            progress,
+            mRenderParameters);
     }
 
-    void UploadHeatBlasterFlame(
+    inline void UploadHeatBlasterFlame(
         vec2f const & centerPosition,
         float radius,
         HeatBlasterActionType action)
@@ -999,12 +808,12 @@ public:
         mFireExtinguisherSprayShaderToRender = Render::ProgramType::FireExtinguisherSpray;
     }
 
-    void UploadShipsStart()
+    inline void UploadShipsStart()
     {
         // Nop
     }
 
-    void UploadShipStart(
+    inline void UploadShipStart(
         ShipId shipId,
         PlaneId maxMaxPlaneId)
     {
@@ -1013,7 +822,7 @@ public:
         mShips[shipId]->UploadStart(maxMaxPlaneId);
     }
 
-    void UploadShipPointImmutableAttributes(
+    inline void UploadShipPointImmutableAttributes(
         ShipId shipId,
         vec2f const * textureCoordinates)
     {
@@ -1022,14 +831,14 @@ public:
         mShips[shipId]->UploadPointImmutableAttributes(textureCoordinates);
     }
 
-    void UploadShipPointMutableAttributesStart(ShipId shipId)
+    inline void UploadShipPointMutableAttributesStart(ShipId shipId)
     {
         assert(shipId >= 0 && shipId < mShips.size());
 
         mShips[shipId]->UploadPointMutableAttributesStart();
     }
 
-    void UploadShipPointMutableAttributes(
+    inline void UploadShipPointMutableAttributes(
         ShipId shipId,
         vec2f const * position,
         float const * light,
@@ -1045,7 +854,7 @@ public:
             lightAndWaterCount);
     }
 
-    void UploadShipPointMutableAttributesPlaneId(
+    inline void UploadShipPointMutableAttributesPlaneId(
         ShipId shipId,
         float const * planeId,
         size_t startDst,
@@ -1059,7 +868,7 @@ public:
             count);
     }
 
-    void UploadShipPointMutableAttributesDecay(
+    inline void UploadShipPointMutableAttributesDecay(
         ShipId shipId,
         float const * decay,
         size_t startDst,
@@ -1073,7 +882,7 @@ public:
             count);
     }
 
-    void UploadShipPointMutableAttributesEnd(ShipId shipId)
+    inline void UploadShipPointMutableAttributesEnd(ShipId shipId)
     {
         assert(shipId >= 0 && shipId < mShips.size());
 
@@ -1082,7 +891,7 @@ public:
 
     // Upload is Asynchronous - buffer may not be used until the
     // next UpdateStart
-    void UploadShipPointColors(
+    inline void UploadShipPointColors(
         ShipId shipId,
         vec4f const * color,
         size_t startDst,
@@ -1103,7 +912,7 @@ public:
 
     // Upload is Asynchronous - buffer may not be used until the
     // next UpdateStart
-    void UploadShipPointTemperature(
+    inline void UploadShipPointTemperature(
         ShipId shipId,
         float const * temperature,
         size_t startDst,
@@ -1221,7 +1030,7 @@ public:
             shipPointIndex2);
     }
 
-    void UploadShipElementStressedSpringsEnd(ShipId shipId)
+    inline void UploadShipElementStressedSpringsEnd(ShipId shipId)
     {
         assert(shipId >= 0 && shipId < mShips.size());
 
@@ -1259,7 +1068,7 @@ public:
             mRenderParameters);
     }
 
-    void UploadShipFlamesEnd(ShipId shipId)
+    inline void UploadShipFlamesEnd(ShipId shipId)
     {
         assert(shipId >= 0 && shipId < mShips.size());
 
@@ -1411,7 +1220,7 @@ public:
             pointIndex);
     }
 
-    void UploadShipElementEphemeralPointsEnd(ShipId shipId)
+    inline void UploadShipElementEphemeralPointsEnd(ShipId shipId)
     {
         assert(shipId >= 0 && shipId < mShips.size());
 
@@ -1438,7 +1247,7 @@ public:
             progress);
     }
 
-    void UploadShipVectors(
+    inline void UploadShipVectors(
         ShipId shipId,
         size_t count,
         vec2f const * position,
@@ -1458,24 +1267,24 @@ public:
             color);
     }
 
-    void UploadShipEnd(ShipId shipId)
+    inline void UploadShipEnd(ShipId shipId)
     {
         assert(shipId >= 0 && shipId < mShips.size());
 
         mShips[shipId]->UploadEnd();
     }
 
-    void UploadShipsEnd()
+    inline void UploadShipsEnd()
     {
         // Nop
     }
 
-    void UploadTextNotificationStart(FontType fontType)
+    inline void UploadTextNotificationStart(FontType fontType)
     {
         mNotificationRenderContext->UploadTextNotificationStart(fontType);
     }
 
-    void UploadTextNotificationLine(
+    inline void UploadTextNotificationLine(
         FontType font,
         std::string const & text,
         AnchorPositionType anchor,
@@ -1490,17 +1299,17 @@ public:
             alpha);
     }
 
-    void UploadTextNotificationEnd(FontType fontType)
+    inline void UploadTextNotificationEnd(FontType fontType)
     {
         mNotificationRenderContext->UploadTextNotificationEnd(fontType);
     }
 
-    void UploadTextureNotificationStart()
+    inline void UploadTextureNotificationStart()
     {
         mNotificationRenderContext->UploadTextureNotificationStart();
     }
 
-    void UploadTextureNotification(
+    inline void UploadTextureNotification(
         TextureFrameId<GenericLinearTextureGroups> const & textureFrameId,
         AnchorPositionType anchor,
         vec2f const & screenOffset, // In texture-size fraction (0.0 -> 1.0)
@@ -1513,7 +1322,7 @@ public:
             alpha);
     }
 
-    void UploadTextureNotificationEnd()
+    inline void UploadTextureNotificationEnd()
     {
         mNotificationRenderContext->UploadTextureNotificationEnd();
     }
@@ -1525,125 +1334,23 @@ public:
     void RenderEnd();
 
 private:
-
-    inline void StoreLightningVertices(
-        float ndcX,
-        float ndcBottomY,
-        float progress,
-        float renderProgress,
-        float personalitySeed,
-        size_t vertexBufferIndex)
-    {
-        if (ndcBottomY > 1.0)
-            return; // Above top, discard
-
-        float constexpr LightningQuadWidth = 0.5f;
-
-        float const leftX = ndcX - LightningQuadWidth / 2.0f;
-        float const rightX = ndcX + LightningQuadWidth / 2.0f;
-        float const topY = 1.0f;
-        float const bottomY = ndcBottomY;
-
-        // Append vertices - two triangles
-
-        // Triangle 1
-
-        // Top-left
-        mLightningVertexBuffer.emplace_at(
-            vertexBufferIndex++,
-            vec2f(leftX, topY),
-            -1.0f,
-            ndcBottomY,
-            progress,
-            renderProgress,
-            personalitySeed);
-
-        // Top-Right
-        mLightningVertexBuffer.emplace_at(
-            vertexBufferIndex++,
-            vec2f(rightX, topY),
-            1.0f,
-            ndcBottomY,
-            progress,
-            renderProgress,
-            personalitySeed);
-
-        // Bottom-left
-        mLightningVertexBuffer.emplace_at(
-            vertexBufferIndex++,
-            vec2f(leftX, bottomY),
-            -1.0f,
-            ndcBottomY,
-            progress,
-            renderProgress,
-            personalitySeed);
-
-        // Triangle 2
-
-        // Top-Right
-        mLightningVertexBuffer.emplace_at(
-            vertexBufferIndex++,
-            vec2f(rightX, topY),
-            1.0f,
-            ndcBottomY,
-            progress,
-            renderProgress,
-            personalitySeed);
-
-        // Bottom-left
-        mLightningVertexBuffer.emplace_at(
-            vertexBufferIndex++,
-            vec2f(leftX, bottomY),
-            -1.0f,
-            ndcBottomY,
-            progress,
-            renderProgress,
-            personalitySeed);
-
-        // Bottom-right
-        mLightningVertexBuffer.emplace_at(
-            vertexBufferIndex++,
-            vec2f(rightX, bottomY),
-            1.0f,
-            ndcBottomY,
-            progress,
-            renderProgress,
-            personalitySeed);
-    }
-
-private:
-
-    void InitializeBuffersAndVAOs();
-    void InitializeCloudTextures(ResourceLocator const & resourceLocator);
-    void InitializeWorldTextures(ResourceLocator const & resourceLocator);
+    
+    void InitializeBuffersAndVAOs();    
+    void InitializeNoiseTextures(ResourceLocator const & resourceLocator);
     void InitializeGenericTextures(ResourceLocator const & resourceLocator);
     void InitializeExplosionTextures(ResourceLocator const & resourceLocator);
 
-    void RenderStars(RenderParameters const & renderParameters);
-    void PrepareRenderLightnings(RenderParameters const & renderParameters);
-    void RenderCloudsAndBackgroundLightnings(RenderParameters const & renderParameters);
-    void RenderOcean(bool opaquely, RenderParameters const & renderParameters);
-    void RenderOceanFloor(RenderParameters const & renderParameters);
-    void RenderAMBombPreImplosions(RenderParameters const & renderParameters);
-    void RenderCrossesOfLight(RenderParameters const & renderParameters);
     void RenderHeatBlasterFlame(RenderParameters const & renderParameters);
     void RenderFireExtinguisherSpray(RenderParameters const & renderParameters);
-    void RenderForegroundLightnings(RenderParameters const & renderParameters);
-    void RenderRain(RenderParameters const & renderParameters);
-    void RenderWorldBorder(RenderParameters const & renderParameters);
 
     void ProcessParameterChanges(RenderParameters const & renderParameters);
     void ApplyViewModelChanges(RenderParameters const & renderParameters);
     void ApplyCanvasSizeChanges(RenderParameters const & renderParameters);
     void ApplyEffectiveAmbientLightIntensityChanges(RenderParameters const & renderParameters);
-    void ApplyOceanDarkeningRateChanges(RenderParameters const & renderParameters);
-    void ApplyOceanRenderParametersChanges(RenderParameters const & renderParameters);
-    void ApplyOceanTextureIndexChanges(RenderParameters const & renderParameters);    
-    void ApplyLandRenderParametersChanges(RenderParameters const & renderParameters);
-    void ApplyLandTextureIndexChanges(RenderParameters const & renderParameters);
 
-    void RecalculateWorldBorder(RenderParameters const & renderParameters);
-    float CalculateEffectiveAmbientLightIntensity() const;
+    static float CalculateEffectiveAmbientLightIntensity(
+        float ambientLightIntensity,
+        float stormAmbientDarkening);
     vec4f CalculateShipWaterColor() const;
 
 private:
@@ -1666,125 +1373,6 @@ private:
 
 #pragma pack(push, 1)
 
-    struct StarVertex
-    {
-        float ndcX;
-        float ndcY;
-        float brightness;
-
-        StarVertex(
-            float _ndcX,
-            float _ndcY,
-            float _brightness)
-            : ndcX(_ndcX)
-            , ndcY(_ndcY)
-            , brightness(_brightness)
-        {}
-    };
-
-    struct LightningVertex
-    {
-        vec2f ndc;
-        float spacePositionX;
-        float ndcBottomY;
-        float progress;
-        float renderProgress;
-        float personalitySeed;
-
-        LightningVertex(
-            vec2f _ndc,
-            float _spacePositionX,
-            float _ndcBottomY,
-            float _progress,
-            float _renderProgress,
-            float _personalitySeed)
-            : ndc(_ndc)
-            , spacePositionX(_spacePositionX)
-            , ndcBottomY(_ndcBottomY)
-            , progress(_progress)
-            , renderProgress(_renderProgress)
-            , personalitySeed(_personalitySeed)
-        {}
-    };
-
-    struct CloudVertex
-    {
-        float ndcX;
-        float ndcY;
-        float ndcTextureX;
-        float ndcTextureY;
-        float darkness;
-
-        CloudVertex(
-            float _ndcX,
-            float _ndcY,
-            float _ndcTextureX,
-            float _ndcTextureY,
-            float _darkness)
-            : ndcX(_ndcX)
-            , ndcY(_ndcY)
-            , ndcTextureX(_ndcTextureX)
-            , ndcTextureY(_ndcTextureY)
-            , darkness(_darkness)
-        {}
-    };
-
-    struct LandSegment
-    {
-        float x1;
-        float y1;
-        float depth1;
-        float x2;
-        float y2;
-        float depth2;
-    };
-
-    struct OceanSegment
-    {
-        float x1;
-        float y1;
-        float value1;
-
-        float x2;
-        float y2;
-        float value2;
-    };
-
-    struct AMBombPreImplosionVertex
-    {
-        vec2f vertex;
-        vec2f centerPosition;
-        float progress;
-        float radius;
-
-        AMBombPreImplosionVertex(
-            vec2f _vertex,
-            vec2f _centerPosition,
-            float _progress,
-            float _radius)
-            : vertex(_vertex)
-            , centerPosition(_centerPosition)
-            , progress(_progress)
-            , radius(_radius)
-        {}
-    };
-
-    struct CrossOfLightVertex
-    {
-        vec2f vertex;
-        vec2f centerPosition;
-        float progress;
-
-        CrossOfLightVertex(
-            vec2f _vertex,
-            vec2f _centerPosition,
-            float _progress)
-            : vertex(_vertex)
-            , centerPosition(_centerPosition)
-            , progress(_progress)
-        {}
-    };
-
     struct HeatBlasterFlameVertex
     {
         vec2f vertexPosition;
@@ -1803,37 +1391,6 @@ private:
         {}
     };
 
-    struct RainVertex
-    {
-        float ndcX;
-        float ndcY;
-
-        RainVertex(
-            float _ndcX,
-            float _ndcY)
-            : ndcX(_ndcX)
-            , ndcY(_ndcY)
-        {}
-    };
-
-    struct WorldBorderVertex
-    {
-        float x;
-        float y;
-        float textureX;
-        float textureY;
-
-        WorldBorderVertex(
-            float _x,
-            float _y,
-            float _textureX,
-            float _textureY)
-            : x(_x)
-            , y(_y)
-            , textureX(_textureX)
-            , textureY(_textureY)
-        {}
-    };
 
 #pragma pack(pop)
 
@@ -1841,82 +1398,22 @@ private:
     // VBOs and uploaded buffers and params
     //
 
-    BoundedVector<StarVertex> mStarVertexBuffer;
-    bool mIsStarVertexBufferDirty;
-    GameOpenGLVBO mStarVBO;
-    size_t mStarVBOAllocatedVertexSize;
-
-    BoundedVector<LightningVertex> mLightningVertexBuffer;
-    size_t mBackgroundLightningVertexCount;
-    size_t mForegroundLightningVertexCount;
-    GameOpenGLVBO mLightningVBO;
-    size_t mLightningVBOAllocatedVertexSize;
-
-    BoundedVector<CloudVertex> mCloudVertexBuffer;
-    GameOpenGLVBO mCloudVBO;
-    size_t mCloudVBOAllocatedVertexSize;
-
-    BoundedVector<LandSegment> mLandSegmentBuffer;
-    GameOpenGLVBO mLandSegmentVBO;
-    size_t mLandSegmentVBOAllocatedVertexSize;
-
-    BoundedVector<OceanSegment> mOceanSegmentBuffer;
-    GameOpenGLVBO mOceanSegmentVBO;
-    size_t mOceanSegmentVBOAllocatedVertexSize;
-
-    std::vector<AMBombPreImplosionVertex> mAMBombPreImplosionVertexBuffer;
-    GameOpenGLVBO mAMBombPreImplosionVBO;
-    size_t mAMBombPreImplosionVBOAllocatedVertexSize;
-
-    std::vector<CrossOfLightVertex> mCrossOfLightVertexBuffer;
-    GameOpenGLVBO mCrossOfLightVBO;
-    size_t mCrossOfLightVBOAllocatedVertexSize;
-
     std::array<HeatBlasterFlameVertex, 6> mHeatBlasterFlameVertexBuffer;
     GameOpenGLVBO mHeatBlasterFlameVBO;
 
     std::array<FireExtinguisherSprayVertex, 6> mFireExtinguisherSprayVertexBuffer;
     GameOpenGLVBO mFireExtinguisherSprayVBO;
 
-    float mStormAmbientDarkening;
-
-    GameOpenGLVBO mRainVBO;
-    float mRainDensity;
-    bool mIsRainDensityDirty;
-
-    std::vector<WorldBorderVertex> mWorldBorderVertexBuffer;
-    GameOpenGLVBO mWorldBorderVBO;
-
     //
     // VAOs
     //
 
-    GameOpenGLVAO mStarVAO;
-    GameOpenGLVAO mLightningVAO;
-    GameOpenGLVAO mCloudVAO;
-    GameOpenGLVAO mLandVAO;
-    GameOpenGLVAO mOceanVAO;
-    GameOpenGLVAO mAMBombPreImplosionVAO;
-    GameOpenGLVAO mCrossOfLightVAO;
     GameOpenGLVAO mHeatBlasterFlameVAO;
     GameOpenGLVAO mFireExtinguisherSprayVAO;
-    GameOpenGLVAO mRainVAO;
-    GameOpenGLVAO mWorldBorderVAO;
 
     //
     // Textures
     //
-
-    GameOpenGLTexture mCloudTextureAtlasOpenGLHandle;
-    std::unique_ptr<TextureAtlasMetadata<CloudTextureGroups>> mCloudTextureAtlasMetadata;
-
-    UploadedTextureManager<WorldTextureGroups> mUploadedWorldTextureManager;
-
-    std::vector<TextureFrameSpecification<WorldTextureGroups>> mOceanTextureFrameSpecifications;
-    GameOpenGLTexture mOceanTextureOpenGLHandle;
-
-    std::vector<TextureFrameSpecification<WorldTextureGroups>> mLandTextureFrameSpecifications;
-    GameOpenGLTexture mLandTextureOpenGLHandle;
 
     GameOpenGLTexture mGenericLinearTextureAtlasOpenGLHandle;
     std::unique_ptr<TextureAtlasMetadata<GenericLinearTextureGroups>> mGenericLinearTextureAtlasMetadata;
@@ -1929,11 +1426,15 @@ private:
 
     UploadedTextureManager<NoiseTextureGroups> mUploadedNoiseTexturesManager;
 
+
     //
-    // Ships
+    // Child contextes
     //
 
     std::vector<std::unique_ptr<ShipRenderContext>> mShips;
+    std::unique_ptr<WorldRenderContext> mWorldRenderContext;
+    std::unique_ptr<NotificationRenderContext> mNotificationRenderContext;
+
 
     //
     // HeatBlaster
@@ -1969,11 +1470,10 @@ private:
     std::function<void()> const mSwapRenderBuffersFunction;
 
     //
-    // Managers
+    // Shader manager
     //
 
-    std::unique_ptr<ShaderManager<ShaderManagerTraits>> mShaderManager;
-    std::unique_ptr<NotificationRenderContext> mNotificationRenderContext;
+    std::unique_ptr<ShaderManager<ShaderManagerTraits>> mShaderManager;    
 
     //
     // Render parameters
